Implemented Rest Controller and Implemented Xhr

+++ src/main/java/ga/alili/controller/StudentControllerRest.java
@@ -32,42 +32,40 @@ public class StudentControllerRest {
         Optional<Student> optionalStudent = studentRepository.getStudentById(id);
         if (optionalStudent.isPresent()){
             Student student = optionalStudent.get();
-            return student;
-        }
-        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Student with id "+ id+ " not found");
+            return student;  }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Student with id " + id + " not found");
     }
     @PostMapping("/list/")
     public Student addStudent(@RequestBody Student student){
-        log.debug("add new student = "+student);
+        log.debug("add new student = " + student);
         student = studentRepository.addStudent(student);
         return student;
     }
-    @PutMapping("/list/")
-    public Student saveStudent(@RequestBody Student student){
+
+    @PutMapping("/list/{id}")
+    public Student saveStudent(@RequestBody Student student, @PathVariable(name = "id") Long id){
         log.debug("insert or update student = "+student);
-        Optional<Student> optionalStudent = studentRepository.getStudentById(student.getId());
+        Optional<Student> optionalStudent = studentRepository.getStudentById(id);
         if (optionalStudent.isPresent()){
             log.debug("student found, update");
-            student = studentRepository.updateStudent(student);
-        } else {
+            student = studentRepository.updateStudent(id, student);
+        }
+        else {
             log.debug("student not found, create new student");
             student = studentRepository.addStudent(student);
         }
-        student = studentRepository.addStudent(student);
         return student;
     }
-//    @DeleteMapping("/list/{id}")
-//    public void deleteStudent(@PathVariable("id") long id){
-//        Optional<Student> optionalStudent = studentRepository.getStudentById(id);
-//        if (optionalStudent.isPresent()){
-//            log.debug("student found, deleting");
-//            studentRepository.deleteStudent(id);
-//        }
-//        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "student with id = "+id+ " not found!");
-//    }
-@DeleteMapping("/list/{id}")
-public void deleteStudent(@PathVariable("id") long id){
-    studentRepository.deleteStudent(id);
 
-}
+    @DeleteMapping("/list/{id}")
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    public void deleteStudent(@PathVariable("id") long id){
+        Optional<Student> optionalStudent = studentRepository.getStudentById(id);
+        if (optionalStudent.isPresent()){
+            log.debug("student found, deleting");
+            boolean result = studentRepository.deleteStudent(id);
+            return;
+        }
+        throw new ResponseStatusException(HttpStatus.NOT_FOUND, "student with id = " + id + " not found!");
+    }
 }
